For tooling, define audio in the V2 location.
This helps as Pyright identifies completions etc. as from the module
where the definition is not the re-export so this helps in the Python
editor.
